requirements.txt
affine==2.4.0
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
cycler==0.11.0
Django==3.2.23
dnspython==2.3.0
docopt==0.6.2
fonttools==4.38.0
GeoAlchemy==0.7.2
GeoAlchemy2==0.14.2
geopandas==0.10.2
geoserver-rest==2.3.4
greenlet==3.0.1
idna==3.4
importlib-metadata==6.7.0
Js2Py==0.74
kiwisolver==1.4.5
matplotlib==3.5.3
munch==4.0.0
packaging==23.2
Pillow==9.5.0
pipwin==0.5.2
postgres-helper==0.0.5
psycopg2==2.9.9
Pygments==2.17.2
pyjsparser==2.7.1
pymongo==4.6.0
pyparsing==3.1.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
Rtree==1.0.1
seaborn==0.12.2
six==1.16.0
snuggs==1.4.7
soupsieve==2.4.1
SQLAlchemy==1.4.46
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
xmltodict==0.13.0
zipp==3.15.0
